cchess/requirements.txt
pyglet==1.4.4
tensorflow==1.14.0
python-chess
